corrections and improvements

+++ atmosphericeffectsdialog.cpp
@@ -13,7 +13,7 @@ AtmosphericEffectsDialog::AtmosphericEffectsDialog(QWidget *parent, CentralWidge
 
     A=laserWindow->getA_Coefficient();
     V=laserWindow->getMeteoRange();
-    atmAttCoeff=3.91/V*powf(550/wavelength, A);
+    atmAttCoeff=3.91/V*std::pow(550/wavelength, A);
     int V_km =V/1000;
     ui->A_Label->setText(QString::number(A, 'e', 2));
     ui->meteoRangeSlider->setValue(V_km);
@@ -30,11 +30,11 @@ void AtmosphericEffectsDialog::on_meteoRangeSlider_valueChanged(int value)
 {
     int V_km=value;
     V=1000*V_km;
-    A=0.06*powf(V, 0.33);
+    A=0.06*std::pow(V, 0.33);
 
     ui->meteoRangeLabel->setText(QString::number(V_km));
     ui->A_Label->setText(QString::number(A, 'e', 2));
-    atmAttCoeff=3.91/(V)*powf(550.0/(wavelength), A);
+    atmAttCoeff=3.91/(V)*std::pow(550.0/(wavelength), A);
     ui->atmosphericCoeffLabel->setText(QString::number(atmAttCoeff, 'e', 2));
 }
